% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrpairs0.R
\docType{data}
\name{lrpairs0}
\alias{lrpairs0}
\title{A dataset storing ligand-receptor pairs for comparisons}
\format{
A data frame with 3239 rows (LR pairs) and 2 columns:
\describe{
  \item{ligand}{ligand name}
  \item{receptor}{receptor name}
}
}
\usage{
lrpairs0
}
\description{
A dataset storing ligand-receptor pairs for comparisons
}
\keyword{datasets}
