% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sclr.R
\name{new_sclr}
\alias{new_sclr}
\alias{is_sclr}
\title{Create a new \code{sclr} object}
\usage{
new_sclr(fit, x, y, cl, mf, mt)

is_sclr(fit)
}
\arguments{
\item{fit}{A list returned by \code{\link{sclr_fit}}.}

\item{x}{Model matrix.}

\item{y}{Model response.}

\item{cl}{Call.}

\item{mf}{Model frame.}

\item{mt}{Model terms.}
}
\value{
\code{sclr} object
}
\description{
\code{new_sclr} creates the object \code{\link{sclr}} returns.
\code{is_sclr} checks if the object is of class \code{sclr}.
}
