% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scLR.R
\name{scLR}
\alias{scLR}
\title{Test dysregulated ligand-receptor interactions at small sample sizes (count matrix)}
\usage{
scLR(
  countmatrix,
  cellinfo,
  lrpairs.sample,
  low.filter = 1,
  p.adjust.method = "BH",
  Brep = 1000,
  Brep0 = 200,
  pv0 = 0.05,
  p.filter = 0.1,
  show.all = FALSE,
  parallel.use = FALSE,
  cpucores = 2,
  zero.impute = TRUE,
  adjust = 3,
  rhos = 0,
  normalization = "ByALL",
  impute.miss.celltype = NA,
  bntest = FALSE,
  sig.bntest = 0.01
)
}
\arguments{
\item{countmatrix}{Input. Count matrix/dataframe with gene names.}

\item{cellinfo}{Input. Information for each cell, including their sampleIDs, conditions, and cell types.}

\item{lrpairs.sample}{Input. A K by 2 matrix/dataframe to show names of ligand-receptor pairs
which will be compared between two conditions.}

\item{low.filter}{Input. It is used to pre-exclude LR pairs with lowly expressed.
The LR pair will be filtered and not be compared if the unnormalized expression mean of the ligand or
the receptor across all samples is less than \emph{low.filter}.
The default is 2. If NULL, no LR pairs are pre-excluded.}

\item{p.adjust.method}{Input (optional). The default is "BH". Other methods include}

\item{Brep}{Input (optional). Number of Monte-Carlo simulation replications. The default is 1000.}

\item{Brep0}{Input (optional). Number of Monte-Carlo simulation replications in the first-stage (rough) selection.
The default is 200. If Brep0 = Brep, no second-stage selection.}

\item{pv0}{Input (optional). Prespecified p-value in the first-stage (rough) selection.
The default is 0.05.}

\item{p.filter}{Input (optional). It is also used to exclude LR pairs with lowly expressed.
If the normalized expression means of the ligand or the receptor in both conditions
are less than the (p.filter)-quantiles of gene means of the corresponding cell type,
the LR pair will be filtered and not be compared. The default is 0.1.}

\item{show.all}{Input (optional). A logical indicator whether or not to show
the results of the ligand-receptor pairs without being compared. The default is FALSE,
showing only the results of the ligand-receptor pairs with being compared.}

\item{parallel.use}{Input (optional). A logical indicator whether or not to use a parallel computation.
The default is FALSE, i.e., a parallel computation will not be used.}

\item{cpucores}{Input (optional). \emph{cpucores} CPU cores (default = 2) are used for parallel computation.}

\item{zero.impute}{Input (optional). A logical indicator whether zero imputation is performed.
If TRUE, when the expressions of ligands or receptors in all samples of one condition are zeros,
zeros will be imputed with a constant of 1 before normalization.}

\item{adjust}{Input (optional). To adjust estimated bandwidth h (bw.SJ * adjust).
The default of \emph{adjust} is 3.}

\item{rhos}{Input (optional). The correlation coefficients in ligand-receptor pairs.
The default is 0, assuming zero correlations for all ligand-receptor pairs.
Users can assume an overall correlation value (between 0 and 1) for all ligand-receptor pairs, e.g, rhos = 0.1;
or set rhos = "est", the Pearson correlation coefficient estimates are used for the significant correlations
(p-values < 0.01 determined by R function cor.test), and 0 for the non-significant correlations.}

\item{normalization}{Input (optional). Perform normalization within each cell type (normalization = "ByCT") or
for all cell types together (normalization = "ByALL"). The default is "ByALL".}

\item{impute.miss.celltype}{Input (optional). Sometimes, some samples do not have specific cell types.
For the missing cell type in the samples, the gene expressions will be replaced with \emph{impute.miss.celltype}.
The default is NA, denoting there not exist the cell types; 0, denoting lowly expressed.}

\item{bntest}{Input (optional). A logical indicator whether or not to perform a bivariate noraml test.
The default is FALSE.}

\item{sig.bntest}{Input (optional). If bntest = TRUE, under what significance level the significant pairs will be further tested
by a bivariate noraml test. The default is 0.05.}
}
\value{
\item{obs.xy.diff}{Observed difference of mean of LR between two conditions (condition 2 minus condition 1)}

\item{null.diff.sd}{Standard deviation of mean difference of LR under null hypothesis}

\item{diff.stat}{obs.xy.diff is divided by null.diff.sd}

\item{pvalue}{P-value}

\item{stage}{Which stage gene pairs pass to}

\item{adj.p}{Adjusted p-value of scLR for multiple comparison}

\item{Welch.t.*}{Statistic, pooled standard error, and p-value of Welch's t-test}

\item{Welch.t.adj.p}{Adjusted p-value of Welch's t-test for multiple comparison}

\item{WRS.*}{Statistic and p-value of Wilcoxon Rank Sum test}

\item{WRS.adj.p}{Adjusted p-value of Wilcoxon Rank Sum test for multiple comparison}

\item{bntest.p}{p-value of bivariate normal test if bntest = TRUE}

\item{bntest.adj.p}{Adjusted p-value of bivariate normal test for multiple comparison if bntest = TRUE}

\item{mu.matrix1}{Sample mean for each gene in condition 1}

\item{mu.matrix2}{Sample mean for each gene in condition 2}

\item{sigma.matrix1}{Posterior residual standard deviation for each gene in condition 1}

\item{sigma.matrix2}{Posterior residual standard deviation for each gene in condition 2}
}
\description{
A function to test dysregulated ligand-receptor interactions from single cell transcriptomics
between two conditions.
}
\examples{
# A simulated data to compare LR pairs between two conditions: TX1 (control) and TX2.
# 3 replicates each condition.
# 1000 genes and 100 cells (5 cell types, 20 cells each) each replicate.
# 10 ligand-receptor gene pairs across 5 cell types are compared between TX1 and TX2.

set.seed(2022)
cellinfo <- data.frame(sampleID = factor(paste0("s", rep(1:6, each=100))),
                      condition = factor(paste0("tx", rep(1:2, each=300))),
                      cellcluster = factor(paste0("cc", rep(rep(1:5, each=20), 6))) )
cellinfo$condition <- relevel(cellinfo$condition, ref="tx1") # Let tx1 be control.
G <- 1000; n <- 600 # 1000 genes, 600 cells
# Data are generated from NB dist.
mu1 <- rgamma(G, shape = 2, rate = 2)
NB_cell <- function(j) rnbinom(G, size = 0.1, mu = mu1)
countmatrix <- as(sapply(1:n, NB_cell),"sparseMatrix"); rownames(countmatrix) <- paste0("g", 1:1000)
# Names of 10 ligand-receptor pairs
lrpairs.sample <- data.frame(ligand=paste0("g", sample(1:1000, 10)),
                            receptor=paste0("g", sample(1:1000, 10)) )
output <- scLR(countmatrix, cellinfo, lrpairs.sample)
head(output$Rs, 20)

# Assume sample s2 does NOT have cell type cc3
idx.remove <- which(cellinfo$sampleID=="s2" & cellinfo$cellcluster=="cc3")
cellinfo1 <- cellinfo[-idx.remove,]; cellinfo1 <- droplevels(cellinfo1)
countmatrix1 <- countmatrix[,-idx.remove]
output <- scLR(countmatrix1, cellinfo1, lrpairs.sample, impute.miss.celltype = NA)
head(output$Rs, 20)
}
