% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scXY2.R
\name{scXY2}
\alias{scXY2}
\title{Test of XY at very small sample sizes (sumExp matrix)}
\usage{
scXY2(
  samplematrix1,
  samplematrix2,
  genepairs,
  No_celltypes,
  Brep = 1000,
  Brep0 = 200,
  pv0 = 0.05,
  p.filter = 0.1,
  show.all = FALSE,
  parallel.use = FALSE,
  cpucores = 2,
  zero.impute = TRUE,
  adjust = 3,
  rhos = 0
)
}
\arguments{
\item{samplematrix1}{Input. Data matrix/dataframe (normal-distribution-like) for condition 1 (genes by replicates * number of cell types).}

\item{samplematrix2}{Input. Data matrix/dataframe (normal-distribution-like) for condition 2 (genes by replicates * number of cell types).}

\item{genepairs}{Input. A K by 4 matrix/dataframe to indicate which K ligand-receptor pairs are compared between two conditions
It may be inputted by either index or name (See the example below for details of input).}

\item{No_celltypes}{Input. Number of cell types.}

\item{Brep}{Input (optional). Number of Monte-Carlo simulation replications. The default is 1000.}

\item{Brep0}{Input (optional). Number of Monte-Carlo simulation replications in the first-stage (rough) selection.
The default is 200. If Brep0 = Brep, no second-stage selection.}

\item{pv0}{Input (optional). Prespecified p-value in the first-stage (rough) selection.
The default is 0.05.}

\item{p.filter}{Input (optional). It is also used to exclude LR pairs with lowly expressed.
If the normalized expression means of the ligand or the receptor in both conditions
are less than the (p.filter)-quantiles of gene means of the corresponding cell type,
the LR pair will be filtered and not be compared. The default is 0.1.}

\item{show.all}{Input (optional). A logical indicator whether or not to show
the results of the ligand-receptor pairs without being compared. The default is FALSE,
showing only the results of the ligand-receptor pairs with being compared.}

\item{parallel.use}{Input (optional). A logical indicator whether or not to use a parallel computation.
The default is FALSE, i.e., a parallel computation will not be used.}

\item{cpucores}{Input (optional). \emph{cpucores} CPU cores (default = 2) are used for parallel computation.}

\item{zero.impute}{Input (optional). A logical indicator whether zero imputation is performed.}

\item{adjust}{Input (optional). To adjust estimated bandwidth h (bw.SJ * adjust).
The default of \emph{adjust} is 3.}

\item{rhos}{Input (optional). The correlation coefficients in ligand-receptor pairs.
The default is 0, assuming zero correlations for all ligand-receptor pairs.
Users can assume an overall correlation value (between 0 and 1) for all ligand-receptor pairs, e.g, rhos = 0.1;
or set rhos = "est", the Pearson correlation coefficient estimates are used for the significant correlations
(p-values < 0.01 determined by R function cor.test), and 0 for the non-significant correlations.}
}
\value{
\item{obs.xy.diff}{Observed difference of mean of XY between two conditions (condition 2 minus condition 1)}

\item{null.diff.sd}{Standard deviation of mean difference of XY under null hypothesis}

\item{diff.stat}{obs.xy.diff is divided by null.diff.sd}

\item{pvalue}{P-value}

\item{stage}{Which stage gene pairs pass to}

\item{Welch.t.*}{Statistic, pooled standard error, and p-value of Welch's t-test}

\item{WRS.*}{Statistic and p-value of Wilcoxon Rank Sum test}

\item{mu.matrix1}{Sample mean for each gene in condition 1}

\item{mu.matrix2}{Sample mean for each gene in condition 2}

\item{sigma.matrix1}{Posterior residual standard deviation for each gene in condition 1}

\item{sigma.matrix2}{Posterior residual standard deviation for each gene in condition 2}
}
\description{
A function to test difference of XY between two samples with very small sizes (sumExp matrix).
}
\examples{
# An example to compare 3 and 5 replicates in TX1 and TX2, respectively.
# 100 genes and 10 cell types for each replicate.
# 5 ligand-receptor gene pairs are compared between replicates in TX1 and TX2.

set.seed(2025)
## Matrix format inputed must be as follows:
samplematrix1 <- cbind(matrix(rnorm(100*10, mean=1.0,sd=0.3),100,10), # replicate 1 (10 cell types)
                       matrix(rnorm(100*10, mean=1.0,sd=0.3),100,10), # replicate 2 (10 cell types)
                       matrix(rnorm(100*10, mean=1.0,sd=0.3),100,10)) # replicate 3 (10 cell types)
samplematrix2 <- cbind(matrix(rnorm(100*10, mean=0.65,sd=0.3),100,10), # replicate 1 (10 cell types)
                       matrix(rnorm(100*10, mean=0.65,sd=0.3),100,10), # replicate 2 (10 cell types)
                       matrix(rnorm(100*10, mean=0.65,sd=0.3),100,10), # replicate 3 (10 cell types)
                       matrix(rnorm(100*10, mean=0.65,sd=0.3),100,10), # replicate 4 (10 cell types)
                       matrix(rnorm(100*10, mean=0.65,sd=0.3),100,10)) # replicate 5 (10 cell types)
No_celltypes <- 10 ## number of cell types

colnames(samplematrix1) <- rep(paste0("cell",1:No_celltypes), 3)
colnames(samplematrix2) <- rep(paste0("cell",1:No_celltypes), 5)
rownames(samplematrix1) <- rownames(samplematrix2) <- paste0("gene",1:nrow(samplematrix1))

## Set low-expressed genes for some cell types of replicates
samplematrix1[82, 4+(0:(3 - 1)) * No_celltypes] <- c(0, 0, 0)
#samplematrix2[82, 4+(0:(5 - 1)) * No_celltypes] <- c(0.01, 0, 0.02, 0.1, 0)

## Set NA (not detected) for some cell types of replicates
samplematrix1[,c(4,12,27,28)] <- NA
samplematrix2[,c(1,13,24,25,33,36,39,47)] <- NA

## lignad-receptor gene pair index
genepairs <- matrix(NA, nrow=5, ncol=4)
colnames(genepairs) <- c("ligand_cell", "ligand_gene",
                         "receptor_cell", "receptor_gene")
genepairs[,c(1,3)] <- sample(1:10, 2*5)
genepairs[,c(2,4)] <- sample(1:100, 2*5)
## output
output1 <- scXY2(samplematrix1, samplematrix2, genepairs, No_celltypes, Brep=1000)
output1$Rs

## Or use ligand-receptor gene pair name
lcname <- c("cell8", "cell4", "cell1", "cell5", "cell10")
lgname <- c("gene85", "gene82", "gene7", "gene94", "gene58")
rcname <- c("cell6", "cell3", "cell2", "cell7", "cell9")
rgname <- c("gene22", "gene57", "gene26", "gene77", "gene14")
pairname <- data.frame(cbind(lcname, lgname, rcname, rgname))
## output
output2 <- scXY2(samplematrix1, samplematrix2, pairname, No_celltypes, Brep=1000)
output2$Rs
}
