% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{tidy.sclr}
\alias{tidy.sclr}
\title{Tidy a \code{sclr} object.}
\usage{
\method{tidy}{sclr}(x, ci_level = 0.95, ...)
}
\arguments{
\item{x}{An object returned by \code{\link{sclr}}.}

\item{ci_level}{Confidence level for the intervals.}

\item{...}{Not used. Needed to match generic signature.}
}
\value{
A \code{\link[tibble]{tibble}} with one row per model parameter. 
Columns:
\item{term}{Name of model parameter.}
\item{estimate}{Point estimate.}
\item{std_error}{Standard error.}
\item{conf_low}{Lower bound of the confidence interval.}
\item{conf_high}{Upper bound of the confidence interval.}
}
\description{
Summarises the objects returned by \code{\link{sclr}} 
into a \code{\link[tibble]{tibble}}.
}
